reviewed and modified
