+++ README.md
@@ -1,7 +1,7 @@
 
 # three-triangle-iterator
 
-**three-triangle-iterator** is a simple utility which make it easy to iterate over all the triangular geometry faces of a [Mesh](https://threejs.org/docs/#api/en/objects/Mesh), Geometry (in version prior to r125 of three.js) or [BufferGeometry](https://threejs.org/docs/#api/en/core/BufferGeometry) objects. It also supports the instanced buffer geometry. This library can work cross platfrom and with your version [three.js](https://threejs.org). This library is backward-compatible with the versions prior to r125 as well as later. We have tested this with specifically r123 and r129 versions of three.js. 
+**three-triangle-iterator** is a simple utility which make it easy to iterate over all the triangular geometry faces of a [Mesh](https://threejs.org/docs/#api/en/objects/Mesh), Geometry (in version prior to r125 of three.js) or [BufferGeometry](https://threejs.org/docs/#api/en/core/BufferGeometry) objects. It also supports the instanced buffer geometry. This library can work cross platfrom and with your version [three.js](https://threejs.org). This library is backward-compatible with the versions prior to r125 or later. 
 
 #### Install
 ```
@@ -40,6 +40,5 @@ forEachTriangle(object, (triangle) => {
 
 ### Author
 
-[Sourabh Soni](https://prolincur.com)\
 [Prolincur Technologies](https://prolincur.com)
