+++ examples/123/package.json
@@ -10,7 +10,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
-    "three": "^0.123.0",
+    "three": "0.123.0",
     "three-triangle-iterator": "^4.2.0",
     "web-vitals": "^1.1.2"
   },
@@ -19,7 +19,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "prettier": "prettier src/**/*.js --write",
+    "prettier": "prettier src/*.{js,jsx} --write",
     "lint": "eslint --fix --ext .js src"
   },
   "devDependencies": {
