+++ src/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Prolincur Technologies LLP.
+ * Copyright (c) 2020-23 Prolincur Technologies LLP.
  * All Rights Reserved.
  */
